// frontend/src/api/generated/model/personRole.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SmartGrao.WebApi | v1
 * OpenAPI spec version: 1.0.0
 */

export type PersonRole = typeof PersonRole[keyof typeof PersonRole];


export const PersonRole = {
  Undefined: 'Undefined',
  Agronomist: 'Agronomist',
  Technician: 'Technician',
  Operator: 'Operator',
  Producer: 'Producer',
  Other: 'Other',
} as const;